+++ tests/__init__.py
@@ -1,0 +1,5 @@
+"""
+POWERJET Auto-Clicker Tests
+
+© 2026 Sergii Sliusar <powerjet777@gmail.com>
+"""